Adicionado camada de serviço para Busca na API
