+++ src/app/app.module.ts
@@ -1,19 +1,20 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { RouteReuseStrategy } from '@angular/router';
+import { NgModule } from "@angular/core";
+import { BrowserModule } from "@angular/platform-browser";
+import { RouteReuseStrategy } from "@angular/router";
+import { HttpClientModule } from "@angular/common/http";
 
-import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
-import { SplashScreen } from '@ionic-native/splash-screen/ngx';
-import { StatusBar } from '@ionic-native/status-bar/ngx';
+import { IonicModule, IonicRouteStrategy } from "@ionic/angular";
+import { SplashScreen } from "@ionic-native/splash-screen/ngx";
+import { StatusBar } from "@ionic-native/status-bar/ngx";
 
-import { AppComponent } from './app.component';
-import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from "./app.component";
+import { AppRoutingModule } from "./app-routing.module";
 
-import { SocialSharing } from '@ionic-native/social-sharing/ngx';
+import { SocialSharing } from "@ionic-native/social-sharing/ngx";
 //import { Clipboard } from '@ionic-native/clipboard/index';
 // import { File } from '@ionic-native/file/ngx';
 
-import { ShareComponent } from './share/share.component';
+import { ShareComponent } from "./share/share.component";
 
 @NgModule({
   declarations: [AppComponent, ShareComponent],
@@ -21,13 +22,14 @@ import { ShareComponent } from './share/share.component';
   imports: [
     BrowserModule,
     IonicModule.forRoot(),
-    AppRoutingModule
+    AppRoutingModule,
+    HttpClientModule
   ],
   providers: [
     StatusBar,
     SplashScreen,
     { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
-    SocialSharing,
+    SocialSharing
     //Clipboard,
     // File
   ],
