+++ src/app/movie.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MovieService } from './movie.service';
+
+describe('MovieService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: MovieService = TestBed.get(MovieService);
+    expect(service).toBeTruthy();
+  });
+});
