+++ src/app/movie.ts
@@ -1,0 +1,17 @@
+export class Movie {
+  id: number;
+  imdb: string;
+  title: string;
+  actors: string;
+  imageUrl: string;
+  imageBigUrl: string;
+  description: string;
+  year: number;
+  quality: string;
+  rating: number;
+  genres: string;
+  items: any;
+  ratingColor: string;
+  runtime: string;
+  trailer: string;
+}
